% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lu.test.R
\name{lu.test}
\alias{lu.test}
\title{One-Sample Lu-Song Test on Variance}
\usage{
lu.test(
  x,
  alternative = "two.sided",
  conf.level = 0.95,
  sigma.squared = 1,
  data.name = NULL,
  R = 1000
)
}
\arguments{
\item{x}{numeric vector of observations. Missing (\code{NA}), undefined (\code{NaN}), and infinite (\code{Inf}, \code{-Inf}) values are allowed but will be removed.}

\item{alternative}{character string indicating the kind of alternative hypothesis. The possible values are \code{"two.sided"} (the default), \code{"greater"}, and \code{"less"}.}

\item{conf.level}{numeric scalar between 0 and 1 indicating the confidence level associated with the confidence interval for the population variance. The default value is \cr \code{conf.level=0.95}}

\item{sigma.squared}{a numeric scalar indicating the hypothesized value of the variance. The default value is \code{sigma.squared=1}.}

\item{data.name}{character string indicating the name of the data used for the test of variance.}

\item{R}{The number of bootstrap replicates to use for bootstrapped statistics.}
}
\value{
A list of class \code{"htest"} containing the results of the hypothesis test.
}
\description{
Estimate the variance, test the null hypothesis using the Lu-Song test that the variance is equal to a user-specified value, and create a bootstrapped confidence interval for the variance.
}
\note{
A one-sample variance test is a type of hypothesis test that is used to determine if the population variability is significantly different from a specific value. The hypothesis test traditionally utilized for these use cases is the chi-squared test on the population variance, which works under the assumption that the underlying population is normally distributed. However, this test is extremely sensitive to departures from normality. This means that for populations that are not normally distributed Type I error rates would be larger than assumed, making a nonparametric test more appropriate. One such alternative hypothesis test that could be used is the Lu-Song test, which is based on a \emph{U}-statistic rather than the chi-squared statistic.

The function \code{lu.test} can be used to perform a hypothesis test for the population variance based on the methodology described by Yuan-yuan Lu and Li-xin Song in their paper "\emph{U}-Statistics Testing Method in Testing Variance Differences of a Single Population". The function itself was adapted from the \code{varTest} function from the package \pkg{EnvStats} for ease of use. The function can then also be used to generate a bootstrapped confidence interval for the population variance as well. The syntax for the bootstrapped confidence interval was adapted from the syntax used in the package \pkg{rcompanion} to again help with ease of use.
}
\section{See Also}{

\link{var.test}.
}

\examples{
# Generate 20 observations from a normal distribution with parameters
# mean=2 and sd=1.  Test the null hypothesis that the true variance is
# equal to 0.5 against the alternative that the true variance is not
# equal to 0.5.
# (Note: the call to set.seed allows you to reproduce this example).

set.seed(23)
dat <- rnorm(20, mean = 2, sd = 1)
lu.test(dat, sigma.squared = 0.5)

#Results of Hypothesis Test
#--------------------------
#
#Null Hypothesis:                 variance = 0.5
#
#Alternative Hypothesis:          True variance is not equal to 0.5
#
#Test Name:                       Lu-Song Test on Variance
#
#Estimated Parameter(s):          variance = 0.753708
#
#Data:                            dat
#
#Test Statistic:                  U-Statistic = 1.7018
#
#P-value:                         0.08879
#
#95\% Confidence Interval:         LCL = 0.4045517
#                                 UCL = 1.0342297

# Note that in this case we would not reject the
# null hypothesis at the 5\% level.

# Clean up
rm(dat)
}
\references{
Lu, Y., Song, L. (2011). \emph{U}-Statistics Testing Method in Testing Variance Differences of a Single Population. \emph{Journal of Jilin University: Science Edition}, 49(6), 1064-1067.

Millard, S. P., Kowarik, A., & Kowarik, M. A. (2022). Package ‘EnvStats’.

Mangiafico, S., & Mangiafico, M. S. (2017). Package ‘rcompanion’. \emph{Cran Repos}, 20, 1-71.
}
\author{
Peter R. Manthey (\email{prjmanthey@gmail.com})
}
\keyword{htest}
\keyword{model}
