% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lusongtest.R
\docType{package}
\name{lusongtest-package}
\alias{lusongtest-package}
\title{lusongtest: Package for Performing a One-Sample Lu-Song Test on Variance}
\description{
An R package for performing Lu-Song tests. This is a nonparametric hypothesis test that performs a one-sample test of variance. The test is based on the methodology described by Yuan-yuan Lu and Li-xin Song in their paper "\emph{U}-Statistics Testing Method in Testing Variance Differences of a Single Population".
}
